refactor: SecurityContextPersistenceInterceptor to be implemented with SecurityContextInterceptor - remove duplicate afterCompletion logic that is controlled by the implemented interface
