+++ src/test/java/nextstep/subway/auth/infrastructure/TokenSecurityContextPersistenceInterceptorTest.java
@@ -16,6 +16,7 @@ import org.springframework.mock.web.MockHttpServletResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.subway.auth.domain.Authentication;
+import nextstep.subway.auth.ui.interceptor.authorization.TokenSecurityContextPersistenceInterceptor;
 import nextstep.subway.member.application.UserDetailService;
 import nextstep.subway.member.domain.LoginMember;
